added optional prop to Textbox, refactored name/id and moved Label to BaseControls
